Fix FAQ build: keep faqs data server-side, split interactive item into client component
